requirements.txt
streamlit==1.31.1
streamlit-authenticator==0.2.3
streamlit-autorefresh==1.0.1

pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2

plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
